!172 LLama model添加GQA功能-前置准备 * 1.arguments 增加GQA参数，与Megatron-LM保持一致

+++ ascendspeed/core/tensor_parallel/layers.py
@@ -235,7 +235,7 @@ class VocabParallelEmbedding(torch.nn.Module):
                                       self.sparse)
         # Mask the output embedding.
         if self.tensor_model_parallel_size > 1:
-            output_parallel *= input_mask[..., None].half()
+            output_parallel *= input_mask[..., None].to(output_parallel.dtype)
         # Reduce across all the model parallel GPUs.
         output = reduce_from_tensor_model_parallel_region(output_parallel)
         if hasattr(self, 'norm'):
